Add co-located functionality for SmartRedis
For some applications, MOM6 might need to be run in the co-located
variant of SmartRedis where the database exists on the nodes of
the simulation. In this case, the model needs to be set on every
database. For now this is done by striding through the ranks
based, short enough so that every node has at least one rank which
calls set_model. This is somewhat fragile, but is a short term
fix; a longer term, more general solution is actively being worked
on.